Transition @kksh/api to use TypeScript completely
without bundling to js at all. There was some problem with bundling with rollup when I use json plugin.

+++ packages/api/build.ts
@@ -1,7 +1,11 @@
+import fs from "fs"
 import { $ } from "bun"
 
-await $`rm -rf dist`
-await $`pnpm build:rollup`
+if (fs.existsSync("dist")) {
+	await $`rm -rf dist`
+}
+fs.mkdirSync("dist")
+// await $`pnpm build:rollup`
 // await $`cp ../schema/manifest-json-schema.json ./dist/schema.json`
 await $`bun ../schema/scripts/print-schema.ts > dist/schema.json`
